% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny.R
\name{get_data}
\alias{get_data}
\title{Get Data from Prometheus}
\usage{
get_data(
  start_time = format(Sys.time() - 60 * 60 * 24 * 7, "\%Y-\%m-\%dT\%H:\%M:\%SZ"),
  end_time = format(Sys.time(), "\%Y-\%m-\%dT\%H:\%M:\%SZ"),
  urlq = prometheus_url[1]
)
}
\arguments{
\item{start_time}{start time in this format: "2025-01-31T11:48:08Z"}

\item{end_time}{end time in this format: "2025-02-07T11:48:33Z"}

\item{urlq}{Prometheus url}
}
\description{
Get Data from Prometheus
}
\examples{
prometheus_url=c("http://pip1.crabdance.com:1507/api/v1/query_range",
                 "http://pip1.crabdance.com:1505/api/v1/query_range")
start_time<-format(Sys.time()-60*60*24*7,"\%Y-\%m-\%dT\%H:\%M:\%SZ")
end_time<-format(Sys.time(),"\%Y-\%m-\%dT\%H:\%M:\%SZ")
filtered_df<-get_data(start_time=start_time,end_time=end_time,urlq=prometheus_url[2])
head(filtered_df)
filtered_df<-get_data(start_time=start_time,end_time=end_time,urlq=prometheus_url[1])
head(filtered_df)
}
\keyword{prometheus}
\keyword{query}
