% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny.R
\name{pimoroni}
\alias{pimoroni}
\title{enviro}
\usage{
pimoroni(prometheus_url, timezone = "Europe/Bucharest")
}
\arguments{
\item{prometheus_url}{prometheus url}

\item{timezone}{the timezone prometheus exports data}
}
\description{
enviro
}
\examples{
pimoroni(prometheus_url=c("http://pip1.crabdance.com:1507/api/v1/query_range",
                          "http://pip1.crabdance.com:1505/api/v1/query_range"))
}
\keyword{enviro}
\keyword{pimoroni}
