% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny.R
\name{plot_gauge}
\alias{plot_gauge}
\title{Plot Gauge}
\usage{
plot_gauge(data = filtered_df, variable = "Temperature")
}
\arguments{
\item{data}{dataframe}

\item{variable}{one of "Temperature" "Humidity" "Pressure" "Lux" "Proximity" "NH3" "Reducing" "Oxidising" "PM1" "PM2.5" "PM10"}
}
\description{
Plot Gauge
}
\examples{
prometheus_url=c("http://pip1.crabdance.com:1507/api/v1/query_range",
                 "http://pip1.crabdance.com:1505/api/v1/query_range")
start_time<-format(Sys.time()-60*60*24*7,"\%Y-\%m-\%dT\%H:\%M:\%SZ")
end_time<-format(Sys.time(),"\%Y-\%m-\%dT\%H:\%M:\%SZ")
filtered_df<-get_data(start_time=start_time,end_time=end_time,urlq=prometheus_url[2])
head(filtered_df)
filtered_df<-get_data(start_time=start_time,end_time=end_time,urlq=prometheus_url[1])
head(filtered_df)
plot_gauge(data=filtered_df,variable="Temperature")
plot_gauge(data=filtered_df,variable="Humidity")
plot_gauge(data=filtered_df,variable="Pressure")
plot_gauge(data=filtered_df,variable="Lux")
plot_gauge(data=filtered_df,variable="Proximity")
plot_gauge(data=filtered_df,variable="NH3")
plot_gauge(data=filtered_df,variable="Reducing")
plot_gauge(data=filtered_df,variable="Oxidising")
plot_gauge(data=filtered_df,variable="PM1")
plot_gauge(data=filtered_df,variable="PM2.5")
plot_gauge(data=filtered_df,variable="PM10")
}
\keyword{gauge}
\keyword{plot}
