% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny.R
\name{plot_time_series}
\alias{plot_time_series}
\title{Plot Time Series}
\usage{
plot_time_series(
  data = filtered_df,
  variable = "Temperature",
  scale = "",
  k = 30
)
}
\arguments{
\item{data}{start time in this format: "2025-01-31T11:48:08Z"}

\item{variable}{one of "Temperature" "Humidity" "Pressure" "Lux" "Proximity" "NH3" "Reducing" "Oxidising" "PM1" "PM2.5" "PM10"}

\item{scale}{scale}

\item{k}{integer width of the rolling window. Must be odd for rollmedian}
}
\description{
Plot Time Series
}
\examples{
prometheus_url=c("http://pip1.crabdance.com:1507/api/v1/query_range",
                 "http://pip1.crabdance.com:1505/api/v1/query_range")
start_time<-format(Sys.time()-60*60*24*7,"\%Y-\%m-\%dT\%H:\%M:\%SZ")
end_time<-format(Sys.time(),"\%Y-\%m-\%dT\%H:\%M:\%SZ")
urlq<-prometheus_url[1]
filtered_df<-get_data(start_time=start_time,end_time=end_time,urlq=urlq)
plot_time_series(data=filtered_df,variable="Temperature")
plot_time_series(data=filtered_df,variable="Humidity")
plot_time_series(data=filtered_df,variable="Pressure")
plot_time_series(data=filtered_df,variable="Lux")
plot_time_series(data=filtered_df,variable="Proximity")
plot_time_series(data=filtered_df,variable="NH3")
plot_time_series(data=filtered_df,variable="Reducing")
plot_time_series(data=filtered_df,variable="Oxidising")
plot_time_series(data=filtered_df,variable="PM1")
plot_time_series(data=filtered_df,variable="PM2.5")
plot_time_series(data=filtered_df,variable="PM10")
}
\keyword{plot}
\keyword{timeseries}
